refactor: button handler using class
Closes #332
